Working on Issues [#31, #36, #37]
