+++ src/main/java/br/com/ppm/commons/collection/CollectionToStringBuilder.java
