+++ src/test/java/br/com/ppm/commons/collection/CollectionToStringBuilderTest.java
